@@ -28,9 +28,16 @@
 	</scm>
 	<properties>
 		<java.version>17</java.version>
-		<spring-boot.version>3.2.4</spring-boot.version>
+		<!-- <spring-boot.version>3.2.4</spring-boot.version> -->
+		<spring-boot.version>3.3.5</spring-boot.version>
 	</properties>
 	<dependencies>
+
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-actuator</artifactId>
+		</dependency>
+
 		<!-- Spring Boot dependencies -->
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
@@ -58,40 +65,68 @@
 		<dependency>
     		<groupId>org.springframework.security</groupId>
     		<artifactId>spring-security-config</artifactId>
-    		<version>6.3.4</version>
+    		<!-- <version>6.3.4</version> -->
 		</dependency>
 		<dependency>
     		<groupId>org.springframework.security</groupId>
     		<artifactId>spring-security-web</artifactId>
-			<version>6.3.4</version>
+			<!-- <version>6.3.4</version> -->
 		</dependency>
+
+
+		<!-- DATABASE STUFF -->
 
 		<!-- Spring Data JPA dependencies -->
     	<dependency>
         	<groupId>jakarta.servlet</groupId>
         	<artifactId>jakarta.servlet-api</artifactId>
-        	<version>5.0.0</version>
+        	 <version>5.0.0</version> 
         	<scope>provided</scope>
     	</dependency>
 		
 		<dependency>
         	<groupId>jakarta.persistence</groupId>
        		 <artifactId>jakarta.persistence-api</artifactId>
-        	<version>3.2.0</version>
-    	</dependency>
+			 <version>3.1.0</version>
+    	</dependency> 
 		
 		<!-- PostgreSQL Driver -->
 		<dependency>
 			<groupId>org.postgresql</groupId>
     		<artifactId>postgresql</artifactId>
-   			<version>42.7.4</version>
 		</dependency>
 
 		<!-- Spring Boot Starter Data JPA (for JPA support) -->
     	<dependency>
        		<groupId>org.springframework.boot</groupId>
         	<artifactId>spring-boot-starter-data-jpa</artifactId>
+
+					<exclusions>
+        				<exclusion>
+           					 <groupId>org.flywaydb</groupId>
+           					 <artifactId>flyway-core</artifactId>
+        					</exclusion>
+    				</exclusions>
     	</dependency>
+
+		 <!-- Optional: Lombok for cleaner code with less boilerplate (e.g., getters/setters) -->
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.18.28</version>
+            <scope>provided</scope>
+        </dependency>
+
+	
+		<!-- DATABASE STUFF end-->
+
+
+		<!-- Core Spring Boot dependency -->
+   		 <dependency>
+       		 <groupId>org.springframework.boot</groupId>
+       		 <artifactId>spring-boot-starter</artifactId>
+   		 </dependency>
+
 
 	</dependencies>
 
