@@ -1,26 +1,45 @@
 package hu.university.etelprojekt.etelprojekt.entity;
 
-import jakarta.persistence.*;
+import java.io.Serializable;
+import java.time.LocalDate;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "payment")
-public class Payment {
+public class Payment implements Serializable {
+
+    private static final long serialVersionUID = 1L;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "payment_id")
     private Long paymentId;
 
+    @ManyToOne
+    @JoinColumn(name = "order_id", nullable = false)
+    private Order order; // Relationship with the Order entity
+
+    @Column(name = "amount", nullable = false)
+    private Double amount;
+
     @Column(name = "payment_method", nullable = false)
     private String paymentMethod;
 
     @Column(name = "payment_date", nullable = false)
-    private String paymentDate;
+    private LocalDate paymentDate;
 
-    @Column(name = "payment_status", nullable = false)
-    private String paymentStatus;
+    @Column(name = "status")
+    private String status;
 
-    // Getters and Setters
+    // Getter and Setter for paymentId
     public Long getPaymentId() {
         return paymentId;
     }
@@ -29,6 +48,25 @@ public class Payment {
         this.paymentId = paymentId;
     }
 
+    // Getter and Setter for order
+    public Order getOrder() {
+        return order;
+    }
+
+    public void setOrder(Order order) {
+        this.order = order;
+    }
+
+    // Getter and Setter for amount
+    public Double getAmount() {
+        return amount;
+    }
+
+    public void setAmount(Double amount) {
+        this.amount = amount;
+    }
+
+    // Getter and Setter for paymentMethod
     public String getPaymentMethod() {
         return paymentMethod;
     }
@@ -37,19 +75,21 @@ public class Payment {
         this.paymentMethod = paymentMethod;
     }
 
-    public String getPaymentDate() {
+    // Getter and Setter for paymentDate
+    public LocalDate getPaymentDate() {
         return paymentDate;
     }
 
-    public void setPaymentDate(String paymentDate) {
+    public void setPaymentDate(LocalDate paymentDate) {
         this.paymentDate = paymentDate;
     }
 
-    public String getPaymentStatus() {
-        return paymentStatus;
+    // Getter and Setter for status
+    public String getStatus() {
+        return status;
     }
 
-    public void setPaymentStatus(String paymentStatus) {
-        this.paymentStatus = paymentStatus;
+    public void setStatus(String status) {
+        this.status = status;
     }
 }
