@@ -1,37 +1,19 @@
 package hu.university.etelprojekt.etelprojekt.repository;
 
 import hu.university.etelprojekt.etelprojekt.entity.Dish;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 @Repository
 public interface DishRepository extends JpaRepository<Dish, Long> {
 
-    // Save or update a dish (CRUD operations are provided by JpaRepository)
-    @SuppressWarnings("unchecked")
-    Dish save(Dish dish);
+    // Find a Dish by its name
+    List<Dish> findByName(String dishName);
 
-    // Find a dish by its ID
-    Optional<Dish> findById(Long dishId);
+    // Find all dishes by a specific menu id
+    List<Dish> findByMenu_Id(Long menuId);
 
-    // Find all dishes in a specific menu
-    List<Dish> findByMenuId(Long menuId);
-
-    // Find dishes by their name (case-insensitive)
-    List<Dish> findByDishNameIgnoreCaseContaining(String dishName);
-
-    // Delete a dish by its ID
-    void deleteById(Long dishId);
-
-    // Custom query to find dishes under a certain price
-    @Query("SELECT d FROM Dish d WHERE d.price < :price")
-    List<Dish> findDishesByPriceLessThan(Double price);
-
-    // Custom query to find dishes by menu ID and price range
-    @Query("SELECT d FROM Dish d WHERE d.menuId = :menuId AND d.price BETWEEN :minPrice AND :maxPrice")
-    List<Dish> findDishesByMenuAndPriceRange(Long menuId, Double minPrice, Double maxPrice);
 }
