@@ -1,34 +1,23 @@
 package hu.university.etelprojekt.etelprojekt.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
 import hu.university.etelprojekt.etelprojekt.entity.Payment;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
+@Repository
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
 
-    // Find a payment by its ID
-    @SuppressWarnings({ "rawtypes", "unchecked" })
-    Optional findById(Long paymentId);
+    Optional<Payment> findByStatus(String status);
 
-    // Find payments by order ID
-    List<Payment> findByOrderId(Long orderId);
+    Optional<Payment> findByOrder_OrderId(Long orderId); // This should now work with 'order' field
 
-    // Find payments by payment method
-    List<Payment> findByPaymentMethod(String paymentMethod);
+    Optional<Payment> findByPaymentMethod(String paymentMethod);
 
-    // Find payments by status
-    List<Payment> findByStatus(String status);
+    Optional<Payment> findByPaymentDate(LocalDate paymentDate); // Updated to LocalDate
 
-    // Custom query to find payments with specific order ID and status
-    @Query("SELECT p FROM Payment p WHERE p.orderId = :orderId AND p.status = :status")
-    List<Payment> findPaymentsByOrderIdAndStatus(@Param("orderId") Long orderId, @Param("status") String status);
-
-    // Custom query to get all payments for a given payment method
-    @Query("SELECT p FROM Payment p WHERE p.paymentMethod = :paymentMethod")
-    List<Payment> findPaymentsByPaymentMethod(@Param("paymentMethod") String paymentMethod);
+    List<Payment> findByAmount(Double amount);
 }
